ヘッダーの色を変更。containerの位置をmarginで調整するように変更
